Cambiado el tipo de dato del modificador de temporada a decimal

+++ database/migrations/2020_12_22_075616_create_temporadas_table.php
@@ -18,7 +18,7 @@ class CreateTemporadasTable extends Migration
             $table->string('nombre');
             $table->date('fecha_inicio');
             $table->date('fecha_fin');
-            $table->integer('mod_temporada');
+            $table->decimal('mod_temporada');
             $table->timestamps();
         });
     }
